Split StoreError into StoreError and AlterTableError
