+++ src/store/mod.rs
@@ -1,29 +1,18 @@
+mod alter_table;
+pub use alter_table::*;
+
 use serde::Serialize;
 use std::fmt::Debug;
 use std::marker::Sized;
 use thiserror::Error;
 
-use sqlparser::ast::ColumnDef;
-
-use super::data::{Row, Schema};
-use super::result::{MutResult, Result};
+use crate::data::{Row, Schema};
+use crate::result::{MutResult, Result};
 
 #[derive(Error, Serialize, Debug, PartialEq)]
 pub enum StoreError {
     #[error("Schema not found")]
     SchemaNotFound,
-
-    // AlterTable error
-    #[error("Column not found")]
-    ColumnNotFound,
-
-    // AlterTable error
-    #[error("Default value is required: {0}")]
-    DefaultValueRequired(String),
-
-    // AlterTable error
-    #[error("Column already exists: {0}")]
-    ColumnAlreadyExists(String),
 }
 
 pub type RowIter<T> = Box<dyn Iterator<Item = Result<(T, Row)>>>;
@@ -50,20 +39,4 @@ where
     fn insert_data(self, key: &T, row: Row) -> MutResult<Self, ()>;
 
     fn delete_data(self, key: &T) -> MutResult<Self, ()>;
-}
-
-pub trait AlterTable
-where
-    Self: Sized,
-{
-    fn rename_schema(self, table_name: &str, new_table_name: &str) -> MutResult<Self, ()>;
-
-    fn rename_column(
-        self,
-        table_name: &str,
-        old_column_name: &str,
-        new_column_name: &str,
-    ) -> MutResult<Self, ()>;
-
-    fn add_column(self, table_name: &str, column_def: &ColumnDef) -> MutResult<Self, ()>;
 }
